+++ uliweb/utils/sorteddict.py
@@ -1,72 +1,73 @@
-class SortedDict(object):
-    def __init__(self, value=None):
-        self._dict = value or {}
-        if value:
-            self._fields = self._dict.keys()
-        else:
-            self._fields = []
+__all__ = ['SortedDict']
+
+from itertools import imap as _imap
+
+class SortedDict(dict):
+    def __init__(self, value=None, kvio=False):
+        """
+        SortedDict will implement Key Insertion Order (KIO: updates of values
+        do not affect the position of the key), Key Value Insertion Order (KVIO,
+        an existing key's position is removed and put at the back)
+        """
+        self._kvio = kvio
+        self._fields = []
+        self.update(value or {})
             
-    def __getitem__(self, key):
-        return self._dict[key]
-    
-    def __getattr__(self, key): 
-        try: 
+    def __setitem__(self, key, value, append=False, dict_setitem=dict.__setitem__):
+        if not key in self:
+            self._fields.append(key)
+        elif self._kvio or append:
+            self._fields.remove(key)
+            self._fields.append(key)
+        return dict_setitem(self, key, value)
+        
+    def __delitem__(self, key, dict_delitem=dict.__delitem__):
+        if key.startswith('_'):
+            del self.__dict__[key]
+        else:
+            dict_delitem(self, key)
+            self._fields.remove(key)
+
+    def __iter__(self):
+        'od.__iter__() <==> iter(od)'
+        for k in self._fields:
+            yield k
+
+    def __reversed__(self):
+        'od.__reversed__() <==> reversed(od)'
+        for k in reversed(self._fields):
+            yield k
+
+    def __getattr__(self, key):
+        try:
             return self.__getitem__(key)
         except KeyError as k:
             return None
-        
-    def __setitem__(self, key, value, append=False):
-        """
-        If append == True, then force existed key append to the end
-        """
-        self._dict[key] = value
-        try:
-            index = self._fields.index(key)
-        except ValueError:
-            index = -1
-        if index > -1:
-            if append:
-                del self._fields[index]
-                self._fields.append(key)
-        else:
-            self._fields.append(key)
-        
+
     def __setattr__(self, key, value):
         if key.startswith('_'):
             self.__dict__[key] = value
         else:
             self.__setitem__(key, value)
-            
-    def __delitem__(self, key):
-        if key.startswith('_'):
-            del self.__dict__[key]
-        else:
-            del self._dict[key]
-            self._fields.remove(key)
-            
+
     def __delattr__(self, key):
-        try: 
+        try:
             self.__delitem__(key)
         except KeyError as k:
             raise AttributeError(k)
-        
-    def __len__(self):
-        return len(self._dict)
-        
-    def __contains__(self, key):
-        return key in self._dict
-    
+
     def keys(self):
         return self._fields
     
     def values(self):
-        return [self._dict[k] for k in self._fields]
+        return [self[k] for k in self._fields]
     
     def iterkeys(self):
-        return self.keys()
+        return iter(self)
     
     def itervalues(self):
-        return self.values()
+        for k in self:
+            yield self[k]
         
     def update(self, value):
         for k, v in value.items():
@@ -76,16 +77,11 @@ class SortedDict(object):
         return [(k, self[k]) for k in self._fields]
     
     def iteritems(self):
-        return self.items()
-    
-    def get(self, key, default=None):
-        try: 
-            return self.__getitem__(key)
-        except KeyError as k:
-            return default
-    
+        for k in self:
+            yield (k, self[k])
+
     def pop(self, key, default=None):
-        v = self._dict.pop(key, default)
+        v = dict.pop(self, key, default)
         if key in self._fields:
             self._fields.remove(key)
         return v
@@ -94,23 +90,35 @@ class SortedDict(object):
         return '<%s {%s}>' % (self.__class__.__name__, ', '.join(['%r:%r' % (k, v) for k, v in sorted(self.items())]))
 
     def dict(self):
-        return self._dict.copy()
+        return self
     
     def copy(self):
-        return self._dict.copy()
+        return self.__class__(self)
     
-    def sort(self, cmp=None, key=None, reverse=False):
-        self._fields = [x for x, y in sorted(self.items(), cmp, key, reverse)]
-        
-    def setdefault(self, key, value):
-        if key in self._dict:
-            return self._dict[key]
+    # def sort(self, cmp=None, key=None, reverse=False):
+    #     self._fields = [x for x, y in sorted(self.items(), cmp, key, reverse)]
+    #
+    def setdefault(self, key, value=None):
+        if key in self:
+            return self[key]
         else:
-            self._dict[key] = value
-            self._fields.append(key)
+            self[key] = value
             return value
 
     def clear(self):
-        self._dict.clear()
+        dict.clear(self)
         self._fields = []
-    
+
+    def __eq__(self, other):
+        '''od.__eq__(y) <==> od==y.  Comparison to another OD is order-sensitive
+        while comparison to a regular mapping is order-insensitive.
+
+        '''
+        if isinstance(other, SortedDict):
+            return dict.__eq__(self, other) and all(_imap(_eq, self, other))
+        return dict.__eq__(self, other)
+
+    def __ne__(self, other):
+        'od.__ne__(y) <==> od!=y'
+        return not self == other
+